requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
albucore=0.0.13=pypi_0
albumentations=1.4.10=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
astor=0.8.1=pypi_0
beautifulsoup4=4.13.3=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2025.1.31=h56e8100_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.12=pypi_0
decorator=5.2.1=pypi_0
dotenv=0.9.9=pypi_0
exceptiongroup=1.2.2=pypi_0
faiss-cpu=1.10.0=pypi_0
fastreid=1.4.0=pypi_0
filelock=3.17.0=pypi_0
fire=0.7.0=pypi_0
fonttools=4.56.0=pypi_0
fsspec=2025.2.0=pypi_0
gdown=5.2.0=pypi_0
grpcio=1.70.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imgaug=0.4.0=pypi_0
jinja2=3.1.5=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.6=h537db12_0
liblzma=5.6.4=h2466b09_0
libsqlite=3.49.1=h67fdade_1
libzlib=1.3.1=h2466b09_2
lmdb=1.6.2=pypi_0
lxml=5.3.1=pypi_0
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.4.2=pypi_0
numpy=1.25.1=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
opencv-python-rolling=5.0.0.20221015=pypi_0
openssl=3.4.1=ha4e3fda_0
opt-einsum=3.3.0=pypi_0
packaging=24.2=pypi_0
paddleocr=2.9.1=pypi_0
paddlepaddle=2.6.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.1.0=pypi_0
pip=25.0.1=pyh8b19718_0
protobuf=3.20.2=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pybaseutils=2.2.19=pypi_0
pyclipper=1.3.0.post6=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pyocr=0.8.5=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.16=h37870fc_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
rapidfuzz=3.12.2=pypi_0
requests=2.32.3=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.8.2=pyhff2d567_0
shapely=2.0.7=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.2.18=pypi_0
tk=8.6.13=h5226925_1
tomli=2.2.1=pypi_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
ultralytics=8.3.82=pypi_0
ultralytics-thop=2.0.14=pypi_0
urllib3=2.3.0=pypi_0
vc=14.3=hbf610ac_24
vc14_runtime=14.42.34438=hfd919c2_24
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
yacs=0.1.8=pypi_0

beautifulsoup4~=4.13.3
tqdm~=4.67.1
ultralytics~=8.3.82
pillow~=11.1.0
torch~=1.13.1
opencv-python~=4.11.0.86
numpy~=1.25.1
pyyaml~=6.0.2
scikit-learn~=1.6.1
paddleocr~=2.9.1
shapely~=2.0.7
requests~=2.32.3
psutil~=7.0.0
matplotlib~=3.10.1
scipy~=1.15.2